# Use variable-size integer for size prefix