Allows strings larger than 65k to be added (assuming the page size is larger) and optimizes storage space for many small strings (<128 bytes).

Broken out from PR #9


Performance impact

Before
| Method | DataSet         | Pool                 | Mean        | Error     | StdDev     | Median      |
|------- |---------------- |--------------------- |------------:|----------:|-----------:|------------:|
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 4,639.18 ns | 92.503 ns | 266.894 ns | 4,563.86 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   401.31 ns |  8.005 ns |   7.862 ns |   399.50 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |   115.02 ns |  0.352 ns |   0.312 ns |   114.98 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 1,533.70 ns | 18.462 ns |  16.366 ns | 1,536.82 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   209.45 ns |  5.817 ns |  16.782 ns |   205.75 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |    90.47 ns |  1.696 ns |   1.503 ns |    90.31 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] |   645.70 ns | 12.502 ns |  14.882 ns |   643.25 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   139.66 ns |  2.829 ns |   7.403 ns |   137.83 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |    61.68 ns |  0.890 ns |   0.789 ns |    61.68 ns |

After
| Method | DataSet         | Pool                 | Mean        | Error     | StdDev     | Median      |
|------- |---------------- |--------------------- |------------:|----------:|-----------:|------------:|
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 4,484.04 ns | 88.240 ns | 174.176 ns | 4,466.85 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   436.64 ns |  8.435 ns |  20.848 ns |   429.44 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |   120.66 ns |  2.289 ns |   2.141 ns |   120.65 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] | 1,601.54 ns | 31.836 ns |  66.454 ns | 1,593.21 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   212.90 ns |  4.962 ns |  14.077 ns |   208.44 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |    90.45 ns |  0.726 ns |   0.644 ns |    90.24 ns |
| DoAdd  | String[1000000] | Utf8S(...)ed=0) [61] |   643.70 ns | 15.213 ns |  43.649 ns |   628.79 ns |
| DoAdd  | String[100000]  | Utf8S(...)ed=0) [61] |   142.94 ns |  7.234 ns |  20.870 ns |   138.05 ns |
| DoAdd  | String[10000]   | Utf8S(...)ed=0) [61] |    65.49 ns |  1.343 ns |   2.169 ns |    65.36 ns |